+++ src/app/auth/[...nextauth]/route.ts
@@ -1,79 +1,7 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-import CredentialsProvider from "next-auth/providers/credentials";
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/auth"; // Import from lib
 
-// Extend NextAuth types to include custom properties
-declare module "next-auth" {
-  interface Session {
-    accessToken?: string;
-  }
+// ✅ Correct App Router pattern (no default export)
+const handler = NextAuth(authOptions);
 
-  interface User {
-    token?: string;
-  }
-}
-
-declare module "next-auth/jwt" {
-  interface JWT {
-    accessToken?: string;
-  }
-}
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    }),
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "text" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        if (!credentials) return null;
-
-        const res = await fetch(`${process.env.BACKEND_URL}/api/auth/login`, {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            email: credentials.email,
-            password: credentials.password,
-          }),
-        });
-
-        const user = await res.json();
-        if (!res.ok) throw new Error(user.message || "Login failed");
-
-        return {
-          id: user.user._id,
-          name: user.user.name,
-          email: user.user.email,
-          token: user.token,
-        };
-      },
-    }),
-  ],
-
-  secret: process.env.NEXTAUTH_SECRET,
-  session: {
-    strategy: "jwt" as const,
-  },
-  pages: { signIn: "/auth/login" },
-
-  callbacks: {
-    // Persist token in JWT
-    async jwt({ token, user }) {
-      if (user) token.accessToken = user.token;
-      return token;
-    },
-    // Expose token in session
-    async session({ session, token }) {
-      session.accessToken = token.accessToken;
-      return session;
-    },
-  },
-};
-
-export default NextAuth(authOptions);
+export { handler as GET, handler as POST };
