+++ src/components/home/CommitmentSection.tsx
@@ -1,56 +1,60 @@
 "use client";
-import { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import Image from "next/image";
 import PageWrapper from "../PageWrapper";
 
 export default function CommitmentSection() {
-  const data = {
-    heading: "COMMITMENT TO COUNTRY DEDICATION TO DUTY",
-    description:
-      "At MailGPT, we are committed to serving with excellence — combining cutting-edge AI, security, and precision in every task we undertake.",
-    features: [
-      { icon: "/home/greenSymbol.png", text: "AI-Powered Inbox Management" },
-      {
-        icon: "/home/greenYellowSymbol.png",
-        text: "Smart Email Categorization",
-      },
-      { icon: "/home/greenSymbol.png", text: "Advanced Spam Protection" },
-      {
-        icon: "/home/greenYellowSymbol.png",
-        text: "Personalized Reply Suggestions",
-      },
-      { icon: "/home/greenSymbol.png", text: "Military-Grade Security" },
-      {
-        icon: "/home/greenYellowSymbol.png",
-        text: "Seamless Multi-Device Sync",
-      },
-    ],
-    progress: [
-      { title: "SECURITY", value: 97, color: "#6B946A" },
-      { title: "EFFICIENCY", value: 92, color: "#997452" },
-      { title: "USER EXPERIENCE", value: 89, color: "#997452" },
-    ],
-    rightImage: "/home/right-image.png",
-    decorImage: "/home/GreenStain.svg",
-  };
+  // ✅ Memoized static data (prevents ESLint dependency warnings)
+  const data = useMemo(
+    () => ({
+      heading: "COMMITMENT TO COUNTRY DEDICATION TO DUTY",
+      description:
+        "At MailGPT, we are committed to serving with excellence — combining cutting-edge AI, security, and precision in every task we undertake.",
+      features: [
+        { icon: "/home/greenSymbol.png", text: "AI-Powered Inbox Management" },
+        {
+          icon: "/home/greenYellowSymbol.png",
+          text: "Smart Email Categorization",
+        },
+        { icon: "/home/greenSymbol.png", text: "Advanced Spam Protection" },
+        {
+          icon: "/home/greenYellowSymbol.png",
+          text: "Personalized Reply Suggestions",
+        },
+        { icon: "/home/greenSymbol.png", text: "Military-Grade Security" },
+        {
+          icon: "/home/greenYellowSymbol.png",
+          text: "Seamless Multi-Device Sync",
+        },
+      ],
+      progress: [
+        { title: "SECURITY", value: 97, color: "#6B946A" },
+        { title: "EFFICIENCY", value: 92, color: "#997452" },
+        { title: "USER EXPERIENCE", value: 89, color: "#997452" },
+      ],
+      rightImage: "/home/right-image.png",
+      decorImage: "/home/GreenStain.svg",
+    }),
+    []
+  );
 
-  // State to trigger animation
+  // ✅ Animated progress bar values
   const [animatedValues, setAnimatedValues] = useState(
     data.progress.map(() => 0)
   );
 
+  // ✅ Trigger animation once on mount
   useEffect(() => {
     const timeout = setTimeout(() => {
       setAnimatedValues(data.progress.map((bar) => bar.value));
-    }, 300); // delay for smooth entry
-
+    }, 300);
     return () => clearTimeout(timeout);
-  }, [0]);
+  }, [data.progress]); // ESLint happy ✅
 
   return (
     <PageWrapper>
       <section className="relative py-12 grid md:grid-cols-2 gap-10 items-center">
-        {/* Decorative Image Outside Left Div */}
+        {/* Decorative Image */}
         <Image
           src={data.decorImage}
           alt="Decorative stain"
@@ -59,17 +63,17 @@ export default function CommitmentSection() {
           className="absolute -top-10 -left-10 z-0 pointer-events-none"
         />
 
-        {/* Left Side */}
+        {/* Left Section */}
         <div className="relative z-10">
           <h2 className="text-2xl md:text-2xl lg:text-3xl font-extrabold leading-snug">
             {data.heading}
           </h2>
 
-          <p className="text-gray-400 text-sm md:text-base max-w-lg">
+          <p className="text-gray-400 text-sm md:text-base max-w-lg mt-2">
             {data.description}
           </p>
 
-          {/* Features */}
+          {/* Features List */}
           <div className="grid grid-cols-2 gap-y-3 mt-6 text-sm md:text-sm">
             {data.features.map((item, idx) => (
               <p key={idx} className="flex items-center gap-2">
@@ -79,7 +83,7 @@ export default function CommitmentSection() {
             ))}
           </div>
 
-          {/* Animated Progress Bars */}
+          {/* Progress Bars */}
           <div className="mt-8 space-y-6 max-w-md">
             {data.progress.map((bar, idx) => (
               <div key={idx}>
@@ -94,14 +98,14 @@ export default function CommitmentSection() {
                       width: `${animatedValues[idx]}%`,
                       backgroundColor: bar.color,
                     }}
-                  ></div>
+                  />
                 </div>
               </div>
             ))}
           </div>
         </div>
 
-        {/* Right Side */}
+        {/* Right Section */}
         <div className="relative flex justify-center md:justify-end">
           <Image
             src={data.rightImage}
